Households data

Household data from the ECV for the years 2005, 2011, and 2019.

households

Format

A tibble (see tibble::tibble-package) with columns:

id_hh

Household identifier.

ecv_year

The year the survey was conducted. Data refer to the previous year.

region

Autonomous Community indicator. Values:

AND	Andalucía

ARA	Aragón

AST	Asturias

BAL	Baleares

CNR	Canarias

CNT	Cantabria

CYL	Castilla y León

CLM	Castilla La Mancha

CAT	Cataluña

VAL	Valencia

EXT	Extremadura

GAL	Galicia

MAD	Madrid

MUR	Murcia

NAV	Navarra

PVA	País Vasco

RIO	La Rioja

CEU	Ceuta

MEL	Melilla

cluster

Primary sampling unit indicator.

weight_hh

Household population weight.

urb

Degree of urbanisation. Factor levels: 'Low', 'Med', and 'High'.

size_hh

Number of persons per household

type_hh

Household type. Factor levels:

HH_01	One member: a man aged less than 30.

HH_02	One member: a man aged between 30 and 64.

HH_03	One member: a man aged 65 or more.

HH_04	One member: a woman aged less than 30.

HH_05	One member: a woman aged between 30 and 64.

HH_06	One member: a woman aged 65 or more.

HH_07	Two adults without dependent children and at least one person aged 65 or more.

HH_08	Two adults without dependent children and both of them aged less than 65.

HH_09	Other households without dependent children.

HH_10	One adult with at least one dependent child.

HH_11	Two adults with one dependent child.

HH_12	Two adults with two dependent children.

HH_13	Two adults with three or more dependent children.

HH_14	Other households with dependent children.

cunits_hh

Consumption units: Modified OECD equivalence scale.

ydisp_hh

Household disposable income.

pov_hh

At risk of poverty indicator.

depriv_hh

Severe material deprivation indicator.

Source

Microdata at INE, the Spanish Statistical Office