Individuals data

Personal data from the ECV for the years 2005, 2011, and 2019.

individuals

Format

A tibble (see tibble::tibble-package) with columns:

id_p

Person identifier.

id_hh

Household identifier.

ecv_year

The year the survey was conducted. Data refer to the previous year.

birth_year

The year of birth.

age

Age, in years.

sex

Factor with levels Female and Male.

absent

Is the individual temporarily absent from the household?

lowjob

Does the individual live in a household with low work intensity?

eu2020

Europe 2020 indicator on poverty and social exclusion.

weight_p16

Weight for individuals over 16 years old.

country_birth

Country of birth. Factor with levels:

Spain	Spain.

EU	European Union (EU 28).

Other	Rest of the world.

nationality

Nationality. Factor with levels: see country_birth.

educ_year

Year in which the individual achieved the highest educational level.

educ

Education level. Factor with levels:

No	Less than primary education.

Prim	Primary education.

LowerSec	Lower secondary education.

UpperSec	Upper secondary education.

Sup	Tertiary education.

work_age

Age at which the individual started working.

work_years

Years working.

full_time

Working full-time.

part_time

Working part-time.

self

Self-employed?

worker

Employee?

unempl

Unemployed?

inactive

Inactive?

has_worked

Has the individual ever worked?

occup

ISCO-08 occupation major group. Factor levels:

ISCO-1	Managers.

ISCO-2	Professionals.

ISCO-3	Technicians and Associate Professionals.

ISCO-4	Clerical Support Workers.

ISCO-5	Services and Sales Workers.

ISCO-6	Skilled Agricultural, Forestry and Fishery Workers.

ISCO-7	Craft and Related Trades Workers.

ISCO-8	Plant and Machine Operators and Assemblers.

ISCO-9	Elementary Occupations.

ISCO-0	Armed Forces Occupations.

Source

Microdata at INE, the Spanish Statistical Office