Intergenerational transmission of poverty data

Data from the ECV for the years 2005, 2011, and 2019. Variables that describe the household of the individuals when they were around 14 years old.

pov_trans

Format

A tibble (see tibble::tibble-package) with columns:

id_p

Person identifier.

ecv_year

The year the survey was conducted. Data refer to the previous year.

age

Age, in years.

mabsent

Mother absent?

fabsent

Father absent?

adults

(Not available in the 2005 survey). Number of adults (18 years old or above).

children

(Not available in the 2005 survey). Number of children (less tha 18 years old).

siblings

(Only in the 2005 survey). Number of siblings.

nworking

Number of people with a job.

fcountry

(Not available in the 2005 survey). Father's country of birth. Factor with levels:

Spain	Spain.

EU	European Union (EU 28).

Other	Rest of the world.

mcountry

(Not available in the 2005 survey). Mother's country of birth. Factor with levels: see fcountry.

fnation

(Not available in the 2005 survey). Father's nationality. Factor with levels: see fcountry.

mnation

(Not available in the 2005 survey). Mother's nationality. Factor with levels: see fcountry.

feduc

Father's education level. Factor levels:

Low	Lower secondary education or less.

Med	Upper secondary education.

Sup	Tertiary education.

meduc

Mother's education level. Factor levels: see feduc

fworker

Was the individual's father self-employed?

mworker

Was the individual's mother self-employed?

fself

Was the individual's father an employee?

mself

Was the individual's mother self-employed?

funempl

Was the individual's father unemployed?

munempl

Was the individual's mother unemployed?

finactive

Was the individual's father inactive?

minactive

Was the individual's mother unemployed?

foccup

ISCO-08 occupation major group for the individual's father. Factor levels:

ISCO-1	Managers.

ISCO-2	Professionals.

ISCO-3	Technicians and Associate Professionals.

ISCO-4	Clerical Support Workers.

ISCO-5	Services and Sales Workers.

ISCO-6	Skilled Agricultural, Forestry and Fishery Workers.

ISCO-7	Craft and Related Trades Workers.

ISCO-8	Plant and Machine Operators and Assemblers.

ISCO-9	Elementary Occupations.

ISCO-0	Armed Forces Occupations.

moccup

ISCO-08 occupation major group for the individual's mother. Factor levels: see foccup.

well_being

(Not available in the 2005 survey). Household well-being when the individual was a teenager. Factor with levels:

Very bad.

Bad.

Fairly bad.

Fairly good.

Good.

Very good.

fin_hardship

(Only in the 2005 survey). Financial difficulties at home when the individual was a teenager. Factor with levels:

Very often.	Often.

Sometimes.	Rarely.

Never.

Source

Microdata at INE, the Spanish Statistical Office