Print method for objects of class "itp"

Prints objects of class "itp" returned from itp.

# S3 method for itp
print(x, all = FALSE, digits = max(3L, getOption("digits") - 3L), ...)

Arguments

x

An object inheriting from class "itp", a result of a call to itp.

all

A logical scalar. If all = FALSE then only the estimated root, the value of the function at the root an the number of iterations are printed. If all = TRUE then, in addition, the final bracketing interval [a, b], the values of the function at the end points of this interval and the estimated precision are printed.

digits

The argument digits passed to format to set the number of significant digits to print.

...

Further arguments to be passed to or from other methods. They are ignored in this function..

Value

The argument x is returned, invisibly.

Details

The default setting is to print only the root, the value of the function at the root and the number of iterations. To include the bracketing interval after convergence and the estimated precision use all = TRUE.

See also